requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
faiss-cpu
pypdf
tiktoken
openai>=1.0.0
pydantic>=2.0.0
python-dotenv
